feat: implement custom hook for registration demo logic
